+++ src/manager/interactive/transform/index.ts
@@ -1,0 +1,4 @@
+export * from "./AbstractTransformManager"
+export * from "./IITranslateManager"
+export * from "./IIResizeManager"
+export * from "./IIRotationManager"